subsystem: factor bus topology out of example system
